Distance check, then move
This checks the distance, then moves the object based on the speed(so
run towards or run away when the player approaches). This should help
you guys get started on your obstacles. Just remember to link the player
to it. I'll upload my worlds later when I have them finished. I just
have to add a few hazards and tweak my enemies.